YINY-54: For all links in the bottom and slide-out menus, the routing mechanism has been replaced. These requests are now routed through routie. These changes break routing for all user-generated content.
I rewrote the front page of the app to use a database lookup to build the page instead of hardcoded values. (The reason I did this is because the
front page was generating links based on old-style query-string-based routes instead of new-style hash-based routes.) In so doing, I fixed a race
condition which affects the app only on the first run after install; localdb now raises an event once it's verified the database installation, and
only then does the app try to build the homepage.